Renamed FE function and moved waitlist functions to their own service file
